Reorganize API methods
