+++ app/Http/Controllers/API/BoothController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\API;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use NotificationChannels\Messagebird\Exceptions\CouldNotSendNotification;
