+++ routes/api.php
@@ -14,21 +14,21 @@ use Illuminate\Http\Request;
 */
 
 /* Booth funcitonality */
-Route::post('/precheck', 'BoothController@precheck');
-Route::post('/request_sms', 'BoothController@requestSms');
-Route::post('/cast_ballot', 'BoothController@castBallot');
+Route::post('/precheck', 'API\BoothController@precheck');
+Route::post('/request_sms', 'API\BoothController@requestSms');
+Route::post('/cast_ballot', 'API\BoothController@castBallot');
 
 /* Front page blocks */
 Route::get('/sidebar', 'HomeController@sidebar');
 Route::get('/option/{option}', 'HomeController@option');
 
 /* Ballot helpers */
-Route::get('/ballot', 'BallotController@ballotJSON');
-Route::get('/ballot/qr/{ref}', 'BallotController@ballotQR');
+Route::get('/ballot', 'API\BallotController@ballotJSON');
+Route::get('/ballot/qr/{ref}', 'API\BallotController@ballotQR');
 
 /* Admin area */
-Route::post('/anull_ballot', 'AdminController@anullBallot')->middleware('auth.api');
-Route::post('/id_lookup', 'AdminController@lookUp')->middleware('auth.api');
-Route::post('/unblock', 'AdminController@unblock')->middleware('auth.api');
-Route::get('/results', 'AdminController@results')->middleware('auth.api');
-Route::get('/reports', 'AdminController@reports')->middleware('auth.api');
+Route::post('/anull_ballot', 'API\AdminController@anullBallot')->middleware('auth.api');
+Route::post('/id_lookup', 'API\AdminController@lookUp')->middleware('auth.api');
+Route::post('/unblock', 'API\AdminController@unblock')->middleware('auth.api');
+Route::get('/results', 'API\AdminController@results')->middleware('auth.api');
+Route::get('/reports', 'API\AdminController@reports')->middleware('auth.api');
